Create README.gallary
Project Description: Image Gallery with Modal View
This project is an Image Gallery with Modal View that allows users to browse a collection of images. When a user clicks on an image, a modal window pops up displaying the image in a larger view. The modal also features "Previous" and "Next" buttons that let users navigate through the gallery seamlessly without closing the modal.